Created basic window/taskpane
I changed the outlook taskpane to include a dropdown for accounts, a dropdown for signatures and a button that opens up a separate popup window which will be used for adding and removing signatures. I would prefer to add a scrolling list for the signature list but this should be a good start.