Gave the 'Edit Signatures...' button very basic functionality
Made it so the 'Edit Signatures...' button in the task pane opens up to a new file '.../src/editsignatures/editsignatures.html'. This file uses the formatting found in .../src/editsignatures/editsignatures.css' which is just a direct pull from the taskpane.css. This page currently has placeholder info on it but shows a proof of concept for any team member wanting to work on this boxes functionality. There is currently no JavaScript functionality